Reorganize files + remove max width
