+++ src/plugins/navigation_embeddable/public/components/editor/navigation_embeddable_editor.scss
@@ -1,4 +1,4 @@
-@import '../mixins';
+@import '../../mixins';
 
 .navEmbeddablePanelEditor {
   max-inline-size: $euiSizeXXL * 18; // 40px * 18 = 720px
